% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueABC.R
\name{trueABC}
\alias{trueABC}
\title{Title}
\usage{
trueABC(
  first_round_ABC_parameters,
  max,
  slope,
  midpoint,
  lysis,
  MOI,
  n_saves,
  n_replicates,
  n_cores,
  accept_crit,
  precent_error,
  write
)
}
\arguments{
\item{first_round_ABC_parameters}{parameter set to start from this can be from the GA or a previous round of ABC}

\item{max}{values from sigmoidal fit of infection data}

\item{slope}{values from sigmoidal fit of infection data}

\item{midpoint}{values from sigmoidal fit of infection data}

\item{lysis}{values from sigmoidal fit of infection data}

\item{MOI}{MOI that experiments were run at}

\item{n_saves}{number of save before the round of ABC is over}

\item{n_replicates}{number of replicate parameter sets to generate}

\item{precent_error}{an exceptable amount of error outside of the bounds of the supplied parameter distributions}

\item{pop_size}{the population size to be used by the GA}

\item{n_core}{number of cores to be used when evaluating final population}

\item{initpop}{if a previous population estimation is to be used to seed arnother round of the GA}
}
\value{
Function returns parameter values accepted from ABC simulation and the max,slope, and midpoint values
}
\description{
Title
}
